backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "r-workspace"
version = "1.2.0"
description = "Accessible multilingual workspace with OAuth2 authentication and OpenAPI documentation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib>=1.7.4,<2.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "sqlalchemy>=2.0.36,<3.0.0",
    "psycopg[binary]>=3.2.3,<4.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
